chech has cycle on undirected graph with uf

+++ src/cplusplus/dsalgo/modular_int_frequent.hpp
@@ -1,4 +1,4 @@
 #include "const_modulus.hpp"
 #include "modular_int.hpp"
-using mint1_000_000_007 = modint<const_mod<int, 1'000'000'007>>;
-using mint998_244_353 = modint<const_mod<int, 998'244'353>>;
+using mint1_000_000_007 = modint<const_mod<long, 1'000'000'007>>;
+using mint998_244_353 = modint<const_mod<long, 998'244'353>>;
